requirements.txt
aiogram==3.4.1        # Telegram-бот
pytube==15.0.0        # скачивание видео с YouTube
yt-dlp==2024.7.16     # скачивание с TikTok, Instagram и других платформ
requests==2.32.3      # работа с API и HTTP-запросами
ffmpeg-python==0.2.0  # обертка над ffmpeg для обработки видео
Pillow==10.2.0        # обработка изображений/оверлеев
moviepy==1.0.3        # простая работа с видео (для фильтров, объединения)
python-dotenv==1.0.1  # хранение токенов в .env
aiofiles==23.2.1      # асинхронная работа с файлами
